# Conflicts:
#	src/main/java/com/nft/app/service/TradingService.java

+++ src/main/java/com/nft/app/entity/TradingDetailsEntity.java
@@ -27,7 +27,7 @@ public class TradingDetailsEntity {
     private Long sellBlockTill;
     private boolean userProfitShared;
     private boolean userProfitBlocked;
-    private Map<String, Long> userProfitBreakup;
+    private Map<String, Double> userProfitBreakup;
     private boolean breakupCreated;
     private TradeStatusEnum tradeStatus;
 }
